# openapi.json
{
  "components": {
    "schemas": {
...
      "ActorResponse": {
        "description": "Response shape for GET /v1/admin/actors/{id} and list.\n\nSecrets (password hashes, token secrets) are never included.\n``oidc_subject`` is omitted from the default view; present in\nthe audit view (``?view=audit``) for traceability.",
        "properties": {
          "_links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          },
          "actor_id": {
            "format": "uuid",
            "title": "Actor Id",
            "type": "string"
          },
          "actor_kind": {
            "title": "Actor Kind",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "oidc_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oidc Subject"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "tenant_id",
          "display_name",
          "email",
          "actor_kind",
          "created_at"
        ],
        "title": "ActorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Links": {
        "description": "HATEOAS-style navigation pointers for a resource.\n\nEvery detail response includes ``_links.self``; richer resources\nexpose pointers to related sub-resources (e.g. capability detail\npoints at its dependencies, artifacts, interface).\n\nURLs preserve the address form the caller used \u2014 slug paths get\nslug URLs back, UUID paths get UUID URLs.",
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "artifacts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts"
          },
          "capability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability"
          },
          "dependencies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencies"
          },
          "interface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interface"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "self": {
            "title": "Self",
            "type": "string"
          },
          "tenant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant"
          }
        },
        "required": [
          "self"
        ],
        "title": "Links",
        "type": "object"
      },
...
      "TenantResponse": {
        "description": "Response shape for GET /v1/admin/tenants/{slug}.\n\nCallers only see their own tenant \u2014 cross-tenant lookup returns 404\nso existence of other tenants is never confirmed through this surface.\n\nAudit view (``?view=audit``) adds ``is_active``; default view omits it\nsince active/inactive is an operator-level detail not needed by\nday-to-day clients.",
        "properties": {
          "_links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "slug",
          "display_name",
          "created_at"
        ],
        "title": "TenantResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "Paste an API token.\n\n- **Local development:** run `make dev-token` \u2014 it seeds a tenant + actor + role grants and prints a fresh token. No env-var setup required if the docker-compose stack is up.\n- **Production / break-glass:** mint with `python scripts/mint_token.py --tenant-id <uuid> --actor-id <uuid> --roles <role>`. Requires the tenant and actor to already exist.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "title": "registry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/actors/{actor_id}": {
      "get": {
        "description": "Return one actor record in the calling tenant.\n\nReturns 404 if the actor does not exist or belongs to a different tenant\nso that cross-tenant actor existence is not confirmed through this surface.\n\nPass ``?view=audit`` to include ``oidc_subject`` in the response.",
        "operationId": "get_actor_v1_admin_actors__actor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Actor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "View"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Actor",
        "tags": [
          "admin: actors"
        ]
      }
    },
...
    "/v1/admin/tenants/{slug}": {
      "get": {
        "description": "Return the calling tenant's record.\n\nThe path accepts the tenant's slug or its UUID in string form.\nReturns 404 if the resolved tenant is not the caller's own tenant \u2014\ncross-tenant lookup is not permitted through this surface (404, not 403,\nso that the existence of other tenants is not confirmed).\n\nPass ``?view=audit`` to include ``is_active`` in the response.",
        "operationId": "get_tenant_v1_admin_tenants__slug__get",
        "parameters": [
          {
            "description": "Tenant slug or UUID string",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Tenant slug or UUID string",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "View"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tenant",
        "tags": [
          "admin: tenants"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}